Upped Spring Boot to 2.15 and updated all other dependencies.
