+++ aspect/pom.xml
@@ -10,18 +10,18 @@
         <version>2.0.0</version>
     </parent>
 
-    <groupId>org.basaki.metrics</groupId>
     <artifactId>metrics-aspect</artifactId>
     <version>2.0.0</version>
-    <name>metrics-aspect</name>
+    <name>Metrics Aspect</name>
 
     <properties>
-        <aspect.version>1.8.10</aspect.version>
+        <aspect.version>1.9.4</aspect.version>
         <java.version>1.8</java.version>
-        <lombok.version>1.16.8</lombok.version>
-        <log4j.version>2.5</log4j.version>
-        <metrics.version>3.2.2</metrics.version>
-        <springboot.version>2.0.4.RELEASE</springboot.version>
+        <junit.version>4.12</junit.version>
+        <lombok.version>1.18.8</lombok.version>
+        <metrics.version>4.1.0</metrics.version>
+        <mockito.version>2.28.2</mockito.version>
+        <springboot.version>2.1.5.RELEASE</springboot.version>
     </properties>
 
     <dependencies>
@@ -48,6 +48,11 @@
         <dependency>
             <groupId>io.dropwizard.metrics</groupId>
             <artifactId>metrics-servlets</artifactId>
+            <version>${metrics.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>io.dropwizard.metrics</groupId>
+            <artifactId>metrics-jmx</artifactId>
             <version>${metrics.version}</version>
         </dependency>
 
@@ -96,12 +101,12 @@
         <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
-            <version>4.12</version>
+            <version>${junit.version}</version>
         </dependency>
         <dependency>
             <groupId>org.mockito</groupId>
             <artifactId>mockito-core</artifactId>
-            <version>1.10.19</version>
+            <version>${mockito.version}</version>
         </dependency>
     </dependencies>
