+++ aspect/src/main/java/org/basaki/metrics/config/JmxMetricsConfiguration.java
@@ -1,8 +1,8 @@
 package org.basaki.metrics.config;
 
-import com.codahale.metrics.JmxReporter;
 import com.codahale.metrics.MetricFilter;
 import com.codahale.metrics.MetricRegistry;
+import com.codahale.metrics.jmx.JmxReporter;
 import java.util.concurrent.TimeUnit;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Qualifier;
