+++ example/src/main/java/org/basaki/example/book/model/Genre.java
@@ -1,0 +1,26 @@
+package org.basaki.example.book.model;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+
+/**
+ * {@code Genre} represents a genre of a {@code Book}.
+ * <p/>
+ *
+ * @author Indra Basak
+ * @since 3/8/17
+ */
+public enum Genre {
+
+    DRAMA, ROMANCE, GUIDE, TRAVEL;
+
+    /**
+     * Returns a <tt>Genre<tt> enum based on string matching
+     *
+     * @param value string stored in database
+     * @return a matching <tt>Genre</tt>
+     */
+    @JsonCreator
+    public static Genre fromValue(String value) {
+        return valueOf(value.toUpperCase());
+    }
+}
